Updated service
added a broadcast reciever to GetLocationService to check if the MapRecordingActivity is running.

if it is running then locations will be broadcasted.
if not they will be stored in an arraylist by the service (in the service).

the MapRecordingActivity will broadcast true in it's onStart
and false in it's onStop